feat: add more style to form and table

+++ northwind/src/app/pages/categories/categories.component.ts
@@ -37,7 +37,6 @@ export class CategoriesComponent implements OnInit {
   setCategorySelected(category:Categories){
     this.categoriesService.setCategorySelected(category)
   }
-  
   getCategories(){
     this.categoriesService.getCategories().subscribe(data=>{
       this.categories = data
@@ -45,9 +44,29 @@ export class CategoriesComponent implements OnInit {
     });
   }
   deleteCategory(id:number){
-    this.categoriesService.deleteCategory(id).subscribe(r=>
-      this.getCategories()
-    );
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.categoriesService.deleteCategory(id).subscribe(r=>{
+          this.getCategories();
+          Swal.fire(
+            'Deleted!',
+            'The category has been deleted.',
+            'success'
+          );
+        }
+        );
+        
+      }
+    })
   }
   
 }
